pyproject.toml
[tool.poetry]
name = "sb1-diploma"
version = "0.1.0"
description = ""
authors = ["Eduard Slobodyanik <slobodyanik.ed@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"
django = "4.2.2"
crispy-bootstrap4 = "^2024.10"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^9.0.1"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
ipython = "^9.0.2"
coverage = "^7.6.12"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"
requests = "^2.32.3"
celery = "^5.4.0"
pydantic = "^2.10.6"
dependency-injector = "^4.46.0"
pony = "^0.7.19"
python-slugify = "^8.0.4"
djangorestframework-camel-case = "^1.4.2"
drf-spectacular = "^0.28.0"
django-extensions = "^3.2.3"
loguru = "^0.7.3"
django-celery-beat = "^2.7.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
